pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pdl-pilot"
version = "0.1.0"
description = "Phase-1 pilot pipeline for dayside magnetosheath Plasma Depletion Layer detection (THEMIS)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
live = ["pyspedas>=1.4", "pytplot>=2.0"]

[project.scripts]
pdl-pilot = "pdl_pilot.cli.run_pilot:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
